# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HEX Sea-Creature League API",
    "version": "1.0.0",
    "description": "Read-only HTTP API over the HEX sea-creature league analysis. Its purpose is to let other applications obtain a staker's sea-creature level and the pool breakdown. The key consumer endpoint is GET /api/whereami. Educational and informational only; see GET /api/disclaimer."
  },
  "servers": [
    { "url": "http://127.0.0.1:3693", "description": "Local dashboard server" }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Liveness and version",
        "responses": {
          "200": {
            "description": "The server is up.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "version": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/whereami": {
      "get": {
        "summary": "Locate a T-Share total (or wallet address set) in the leagues",
        "description": "Given a T-Share total (`tshares`) or one or more wallet addresses (`address`, repeatable, summed), returns the sea-creature level, rank, percentile, and distance to the next tier for each view (eth, pls, combined). This is the endpoint other apps call for the sea-creature level.",
        "parameters": [
          {
            "name": "tshares",
            "in": "query",
            "required": false,
            "schema": { "type": "string" },
            "description": "Total HEX Staking Pool T-Shares (a number, up to 12 decimals)."
          },
          {
            "name": "address",
            "in": "query",
            "required": false,
            "schema": { "type": "array", "items": { "type": "string" } },
            "style": "form",
            "explode": true,
            "description": "One or more EVM addresses (summed). Ignored if `tshares` is given."
          }
        ],
        "responses": {
          "200": {
            "description": "The location in each view.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WhereAmI" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/what-is-my-hex-staking-sea-creature": {
      "get": {
        "summary": "The caller's sea-creature ranking plus the full pool data",
        "description": "Given the caller's total HEX Staking Pool T-Share count as a non-zero positive number with up to 8 decimal places (`tshares`), returns their sea-creature ranking in each view (eth, pls, combined) and the whole staking-pool sub-object: per-chain as-of block + UTC time, pinned-tip chain state, per-view pool totals and tier breakdown, and the sea-creature roster as a key-value map. Every T-Share figure in the response is given to 8 decimal places (T-Shares can be very expensive). Read-only over the cached report; never triggers a scan.",
        "parameters": [
          {
            "name": "tshares",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "pattern": "^\\d+(\\.\\d{1,8})?$" },
            "description": "Total HEX Staking Pool T-Shares: a non-zero positive number, up to 8 decimal places."
          }
        ],
        "responses": {
          "200": {
            "description": "The ranking and the full pool data.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatIsMySeaCreature"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/get-hex-staking-pool-sea-creature-data": {
      "get": {
        "summary": "The full staking-pool data (no parameter)",
        "description": "Everything /api/what-is-my-hex-staking-sea-creature returns except the per-caller ranking: per-chain as-of block + UTC time, pinned-tip chain state, per-view pool totals and tier breakdown, and the sea-creature roster as a key-value map. Every T-Share figure is given to 8 decimal places. Read-only over the cached report; never triggers a scan.",
        "responses": {
          "200": {
            "description": "The pool data.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PoolDataResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/summary": {
      "get": {
        "summary": "The full report summary (all three views)",
        "description": "The machine-readable report: per-view pool totals, tier buckets, the non-OA ranking, pre-formatted display strings, and per-chain as-of tip data.",
        "responses": {
          "200": {
            "description": "The summary object.",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/status": {
      "get": {
        "summary": "Data freshness and update progress",
        "responses": {
          "200": {
            "description": "The current status.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status" }
              }
            }
          }
        }
      }
    },
    "/api/update": {
      "post": {
        "summary": "Start a scan (initial or incremental)",
        "description": "Runs the scan -> oa -> report pipeline in the background: the initial full scan, or an incremental top-up when the cached data is from a prior UTC day. Poll GET /api/status for progress; POST /api/update/stop cancels a run.",
        "responses": {
          "202": {
            "description": "The update started.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "started": { "type": "boolean" } }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Error" },
          "405": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/update/stop": {
      "post": {
        "summary": "Stop the running scan",
        "description": "Cooperatively cancels the in-progress update. The pipeline halts at the next checkpoint-safe boundary, so the cache stays resumable and a later POST /api/update tops up from where it stopped. The dashboard's header Sync button toggles to \"Stop Sync\" while a scan runs and calls this.",
        "responses": {
          "202": {
            "description": "The stop was requested.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "stopping": { "type": "boolean" } }
                }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Error" },
          "405": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/disclaimer": {
      "get": {
        "summary": "The canonical disclaimer text",
        "responses": {
          "200": {
            "description": "The disclaimer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "disclaimer": { "type": "string" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "summary": "This OpenAPI document (self-describing API)",
        "responses": {
          "200": {
            "description": "The OpenAPI spec.",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Error": {
        "description": "An error.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": { "error": { "type": "string" } }
            }
          }
        }
      }
    },
    "schemas": {
      "League": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "emoji": { "type": "string" },
          "name": { "type": "string" }
        }
      },
      "ViewResult": {
        "type": "object",
        "properties": {
          "input": { "type": "string", "enum": ["tshares", "address"] },
          "tshares": { "type": "string" },
          "league": { "$ref": "#/components/schemas/League" },
          "rank": { "type": "integer" },
          "nonOaStakerCount": { "type": "integer" },
          "percentile": { "type": "number" },
          "nextTier": {
            "oneOf": [
              { "$ref": "#/components/schemas/League" },
              { "type": "null" }
            ]
          },
          "distanceToNextTier": { "type": "string" },
          "oaHits": { "type": "array", "items": { "type": "string" } },
          "found": { "type": "array", "items": { "type": "string" } },
          "missing": { "type": "array", "items": { "type": "string" } }
        }
      },
      "WhereAmI": {
        "type": "object",
        "properties": {
          "disclaimer": { "type": "string" },
          "results": {
            "type": "object",
            "properties": {
              "eth": { "$ref": "#/components/schemas/ViewResult" },
              "pls": { "$ref": "#/components/schemas/ViewResult" },
              "combined": { "$ref": "#/components/schemas/ViewResult" }
            }
          }
        }
      },
      "SharePair": {
        "type": "object",
        "properties": {
          "shares": {
            "type": "string",
            "description": "Raw shares (10^12 raw = 1 T-Share)."
          },
          "tShares": {
            "type": "string",
            "description": "The same value in T-Shares, to 8 decimal places."
          }
        }
      },
      "AsOfChain": {
        "type": "object",
        "description": "The last block scanned for a chain and that block's UTC time.",
        "properties": {
          "block": { "type": ["integer", "null"] },
          "timeUtc": {
            "type": ["string", "null"],
            "description": "ISO 8601 UTC timestamp of the block."
          }
        }
      },
      "RosterEntry": {
        "type": "object",
        "properties": {
          "emoji": { "type": "string" },
          "name": { "type": "string" },
          "rank": {
            "type": ["integer", "null"],
            "description": "1 = top tier (Poseidon); null for Plankton."
          },
          "minShareFraction": {
            "type": "string",
            "description": "Minimum share of the non-OA pool, exact fraction (e.g. 1/10)."
          },
          "minSharePercent": {
            "type": "string",
            "description": "The same threshold as a percent (e.g. 10)."
          },
          "note": { "type": "string" }
        }
      },
      "TierData": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "emoji": { "type": "string" },
          "name": { "type": "string" },
          "minShares": { "type": "string" },
          "minTShares": { "type": "string" },
          "wallets": { "type": "integer" },
          "cumulativeWallets": { "type": "integer" },
          "tierShares": { "type": "string" },
          "tierTShares": { "type": "string" }
        }
      },
      "ViewData": {
        "type": "object",
        "properties": {
          "pool": {
            "type": "object",
            "properties": {
              "total": { "$ref": "#/components/schemas/SharePair" },
              "nonOa": { "$ref": "#/components/schemas/SharePair" },
              "oaExcluded": { "$ref": "#/components/schemas/SharePair" }
            }
          },
          "stakerCounts": {
            "type": "object",
            "properties": {
              "nonOa": { "type": "integer" },
              "oa": { "type": "integer" }
            }
          },
          "tiers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TierData" }
          }
        }
      },
      "PoolData": {
        "type": "object",
        "description": "The whole staking-pool picture the app holds, pure over the cached report.",
        "properties": {
          "asOf": {
            "type": "object",
            "properties": {
              "eth": { "$ref": "#/components/schemas/AsOfChain" },
              "pls": { "$ref": "#/components/schemas/AsOfChain" }
            }
          },
          "chains": {
            "type": "object",
            "description": "Pinned-tip chain state per chain (block, timeUtc, stakeSharesTotal, nextStakeSharesTotal, shareRate, currentDay, totalSupply)."
          },
          "views": {
            "type": "object",
            "properties": {
              "eth": { "$ref": "#/components/schemas/ViewData" },
              "pls": { "$ref": "#/components/schemas/ViewData" },
              "combined": { "$ref": "#/components/schemas/ViewData" }
            }
          },
          "roster": {
            "type": "object",
            "description": "Sea-creature roster keyed by tier id (poseidon … plankton).",
            "additionalProperties": {
              "$ref": "#/components/schemas/RosterEntry"
            }
          }
        }
      },
      "PoolDataResponse": {
        "type": "object",
        "properties": {
          "disclaimer": { "type": "string" },
          "stakingPool": { "$ref": "#/components/schemas/PoolData" }
        }
      },
      "WhatIsMySeaCreature": {
        "type": "object",
        "properties": {
          "disclaimer": { "type": "string" },
          "query": {
            "type": "object",
            "properties": { "tshares": { "type": "string" } }
          },
          "ranking": {
            "type": "object",
            "properties": {
              "eth": { "$ref": "#/components/schemas/ViewResult" },
              "pls": { "$ref": "#/components/schemas/ViewResult" },
              "combined": { "$ref": "#/components/schemas/ViewResult" }
            }
          },
          "stakingPool": { "$ref": "#/components/schemas/PoolData" }
        }
      },
      "Status": {
        "type": "object",
        "properties": {
          "hasCompleteScan": { "type": "boolean" },
          "ageHours": { "type": ["number", "null"] },
          "updateEnabled": { "type": "boolean" },
          "reason": { "type": "string" },
          "updating": { "type": "boolean" },
          "progress": { "type": "number" },
          "phase": { "type": "string" }
        }
      }
    }
  }
}